fix: MoveNet input type correction (INT32 → UINT8)
Phase 2 완성:
- MoveNet Thunder 모델 다운로드 완료 (12MB)
- TensorFlow 2.15.0 + TensorFlow Hub 설치
- movenet_analyzer.py 입력 타입 수정
- 테스트 성공 (Confidence: 21%, 6/17 keypoints)

Dependencies:
- tensorflow==2.15.0
- tensorflow-hub
- opencv-python==4.10.0.84
- numpy==1.26.4 (TensorFlow 호환)

✅ MoveNet 실행 준비 완료!

+++ src/Multi/version3/analysis/movenet_analyzer.py
@@ -129,8 +129,8 @@ class MoveNetAnalyzer:
         # 전처리: 256x256 리사이즈
         img_resized = cv2.resize(img_rgb, (self.input_size, self.input_size))
 
-        # Float32로 변환 (0~255 → 0~255, MoveNet은 정규화 필요 없음)
-        img_input = np.expand_dims(img_resized, axis=0).astype(np.int32)
+        # UINT8로 변환 (0~255, MoveNet은 정규화 필요 없음)
+        img_input = np.expand_dims(img_resized, axis=0).astype(np.uint8)
 
         # 추론
         self.interpreter.set_tensor(self.input_details[0]['index'], img_input)
